pyproject.toml
[project]
name="dosagelimitsindex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python=">=3.9"
dependencies = [
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
]
